style(ext-events): split errorMonitor dispatch into a sibling module (2000-line cap)
